requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic-settings
passlib[bcrypt]
python-multipart
openai
chromadb
